Move charts tests to their own module
